+++ src/Settings/SettingsView.tsx
@@ -33,12 +33,12 @@ function SettingsDialog(props: Props) {
   };
 
   return (
-      <div data-testid="settings-pane" style={{ height: '100%', display: 'flex' }}>
-        {/* Outer box containing left-hand fixed-width column with setting sub-categories, and right-hand adjustable width colum with selected subcategory settings. Force height to 100% so that the left hand list border always stretches from top to bottom */}
-        <Box id="test" sx={{ display: 'flex', flexDirection: 'row' }}>
+      <div data-testid="settings-pane" style={{ width: '100%', height: '100%', display: 'flex', flexGrow: 1, flexDirection: 'row', overflowY: 'hidden' }}>
+          {/* Outer box containing left-hand fixed-width column with setting sub-categories, and right-hand adjustable width colum with selected subcategory settings. Force height to 100% so that the left hand list border always stretches from top to bottom */}
           {/* Add a little border to the right-hand side to separate from sub-category settings */}
-          <Box
-            sx={{
+          <div
+            id="settings-pane-left"
+            style={{
               minWidth: '180px',
               marginRight: '10px',
               borderRight: 1,
@@ -88,16 +88,15 @@ function SettingsDialog(props: Props) {
                 </ListItemButton>
               </List>
             </nav>
-          </Box>
+          </div>
           {/* Container to wrap scrollable content in right-hand side pane */}
-          <Box sx={{ overflowY: 'auto' }}>
+          <div style={{ flexGrow: 1, overflowY: 'auto', height: '100%' }}>
             {
               displayedSettingsCategory[
                 appStore.settings.activeSettingsCategory
               ]
             }
-          </Box>
-        </Box>
+          </div>
       </div>
   );
 }